[fix]fix the bug of server
